# docs/api-reference/openapi-file-service.yaml
openapi: 3.0.3
info:
  title: ClawTeam File Service API
  version: "2026-03-23"
  description: |
    File Service contract for /api/v1/files.
    This contract reflects current implementation in packages/api/src/file-service/routes.ts.
servers:
  - url: http://localhost:3000/api/v1
    description: Local API server
security:
  - bearerAuth: []
...
paths:
  /files:
    get:
      tags: [Files]
      summary: List file nodes
      parameters:
        - in: query
          name: parentId
          schema: { type: string, format: uuid }
        - in: query
          name: scope
          schema:
            $ref: '#/components/schemas/FileScope'
        - in: query
          name: scopeRef
          schema: { type: string, format: uuid }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: File list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /files/docs:
    post:
      tags: [Docs]
      summary: Create doc
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string }
                content: { type: string }
                parentId: { type: string, format: uuid }
                scope:
                  $ref: '#/components/schemas/FileScope'
                scopeRef: { type: string, format: uuid }
                taskId: { type: string, format: uuid }
                metadata: { type: object, additionalProperties: true }
                clientToken: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
  schemas:
    FileScope:
      type: string
      enum: [bot_private, task, team_shared]
...
    SuccessResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object
          additionalProperties: true
        traceId:
          type: string
...